requirements.txt
django-braces
django-taggit
youtube-dl
